requirements.txt
pyxdg~=0.26
tasklib>=1.3.0,<2.0.0
filemagic~=1.6
mime_parser~=1.2.0
